Create new /assay/index.ts file and dir and move assay related app components
