+++ packages/components/src/assay/AssayTypeSummary.tsx
@@ -1,9 +1,9 @@
 import React, { FC, memo, useState } from 'react';
 
-import { SelectView, SelectViewInput } from '../base/SelectViewInput';
-import { AppURL } from '../../url/AppURL';
+import { SelectView, SelectViewInput } from '../internal/components/base/SelectViewInput';
+import { AppURL } from '../internal/url/AppURL';
 
-import { StatusGrid } from './StatusGrid';
+import { StatusGrid } from '../internal/components/assay/StatusGrid';
 import { AssaysHeatMap } from './AssaysHeatMap';
 
 const ASSAY_VIEWS = [SelectView.Grid, SelectView.Heatmap];
