+++ packages/components/src/entities/index.ts
@@ -46,6 +46,8 @@ import { SamplesTabbedGridPanel } from './SamplesTabbedGridPanel';
 import { SampleTypeTemplateDownloadRenderer, downloadSampleTypeTemplate } from './SampleTypeTemplateDownloadRenderer';
 import { SampleTypePage } from './SampleTypePage';
 import { SampleIndexNav, SampleTypeIndexNav } from './SampleNav';
+import { SamplesResolver } from './SamplesResolver';
+import { AssayImportSubMenuItem } from './AssayImportSubMenuItem';
 
 export {
     PICKLIST_SAMPLES_FILTER,
@@ -60,6 +62,7 @@ export {
     getSampleWizardURL,
     isFindByIdsSchema,
     loadSampleTypes,
+    AssayImportSubMenuItem,
     CreateSamplesSubMenu,
     CreateSamplesSubMenuBase,
     EntityCrossProjectSelectionConfirmModal,
@@ -87,6 +90,7 @@ export {
     SampleFinderSection,
     SampleIndexNav,
     SampleLineageGraph,
+    SamplesResolver,
     SampleSetDeleteModal,
     SampleTimelinePageBase,
     SampleTypeIndexNav,
